requirements.txt
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
fonttools==4.53.1
GPy==1.13.2
graphviz==0.20.3
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.2
neat-python==0.92
numpy==1.26.4
packaging==24.1
paramz==0.9.6
pillow==10.4.0
pybullet==3.2.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.5.1
scipy==1.12.0
six==1.16.0
threadpoolctl==3.5.0
